🚚 api: get-certification-attestation (first route) : move
controller to certification course src
