+++ api/lib/application/certifications/certification-controller.js
@@ -2,7 +2,6 @@ import { usecases } from '../../domain/usecases/index.js';
 import * as events from '../../domain/events/index.js';
 import * as privateCertificateSerializer from '../../infrastructure/serializers/jsonapi/private-certificate-serializer.js';
 import * as shareableCertificateSerializer from '../../infrastructure/serializers/jsonapi/shareable-certificate-serializer.js';
-import * as certificationAttestationPdf from '../../infrastructure/utils/pdf/certification-attestation-pdf.js';
 import * as requestResponseUtils from '../../infrastructure/utils/request-response-utils.js';
 
 const findUserCertifications = async function (request) {
@@ -31,29 +30,6 @@ const getCertificationByVerificationCode = async function (request, h, dependenc
 
   const shareableCertificate = await usecases.getShareableCertificate({ verificationCode, locale });
   return shareableCertificateSerializer.serialize(shareableCertificate);
-};
-
-const getPDFAttestation = async function (request, h, dependencies = { certificationAttestationPdf }) {
-  const userId = request.auth.credentials.userId;
-  const certificationId = request.params.id;
-  const { i18n } = request;
-  const { isFrenchDomainExtension } = request.query;
-
-  const attestation = await usecases.getCertificationAttestation({
-    userId,
-    certificationId,
-  });
-
-  const { buffer, fileName } = await dependencies.certificationAttestationPdf.getCertificationAttestationsPdfBuffer({
-    certificates: [attestation],
-    isFrenchDomainExtension,
-    i18n,
-  });
-
-  return h
-    .response(buffer)
-    .header('Content-Disposition', `attachment; filename=${fileName}`)
-    .header('Content-Type', 'application/pdf');
 };
 
 const neutralizeChallenge = async function (request, h, dependencies = { events }) {
@@ -86,7 +62,6 @@ const certificationController = {
   findUserCertifications,
   getCertification,
   getCertificationByVerificationCode,
-  getPDFAttestation,
   neutralizeChallenge,
   deneutralizeChallenge,
 };
